It was exposed some interesting info in the webapi project, and a small refactoring
