+++ MartianRobotsSimulation/Application/IRobotProcessor.cs
@@ -1,4 +1,4 @@
-﻿using Domain;
+﻿using Domain.Interfaces;
 using System.Collections.Generic;
 
 namespace Application
