+++ MartianRobotsSimulation/Application/Position/NorthPosition.cs
@@ -1,4 +1,4 @@
-﻿using Domain;
+﻿using Domain.Interfaces;
 
 namespace Application.Position
 {
