+++ MartianRobotsSimulation/Domain/Interfaces/IMisionRepository.cs
@@ -1,12 +1,13 @@
 ﻿using Domain.Models;
 using System.Collections.Generic;
 
-namespace Domain
+namespace Domain.Interfaces
 {
     public interface IMisionRepository
     {
         List<MisionModel> GetAllMisions();
         MisionModel GetLastMision();
+        MisionModel GetById(int id);
         void Add(MisionModel misionModel);
     }
 }
