+++ MartianRobotsSimulation/Domain/Interfaces/IMisionResult.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Domain
+﻿namespace Domain.Interfaces
 {
     public interface IMisionResult
     {
