+++ MartianRobotsSimulation/Domain/Interfaces/ISurface.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Domain
+namespace Domain.Interfaces
 {
     public interface ISurface
     {
