+++ MartianRobotsSimulation/Domain/Scent.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using Domain.Interfaces;
 
 namespace Domain
 {
