+++ MartianRobotsSimulation/Infrastructure/Injection.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using Domain.Interfaces;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.EntityFrameworkCore;
-using Domain;
 
 namespace Infrastructure
 {
