+++ MartianRobotsSimulation/MartianRobotsApi/MartianRobotsApi.csproj
@@ -23,7 +23,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Application\Application.csproj" />
-    <ProjectReference Include="..\Infrastructure\Infrastructure.csproj" />
   </ItemGroup>
   
   <PropertyGroup>
